+++ src/api/v2/orders.rs
@@ -69,10 +69,7 @@ Endpoint! {
   }
 
   fn query(input: &Self::Input) -> Result<Option<Str>, Self::ConversionError> {
-    // TODO: Realistically there should be no way for this unwrap to
-    //       ever panic because our conversion to strings should not be
-    //       fallible. But still, ideally we would not have to unwrap.
-    Ok(Some(to_query(input).unwrap().into()))
+    Ok(Some(to_query(input)?.into()))
   }
 }